+++ package.json
@@ -23,7 +23,8 @@
     "mongoose": "^6.1.6",
     "mongoose-paginate-v2": "^1.5.0",
     "morgan": "^1.10.0",
-    "passport": "^0.5.2"
+    "passport": "^0.5.2",
+    "yup": "^0.32.11"
   },
   "devDependencies": {
     "@babel/cli": "^7.16.8",
